add presentator+ linter
