+++ src/utils/film.js
@@ -1,18 +1,16 @@
+// const getCurrentDate = () => {
+//   const currentDate = new Date();
+//   // currentDate.setHours(23, 59, 59, 999);
 
-
-const getCurrentDate = () => {
-  const currentDate = new Date();
-  currentDate.setHours(23, 59, 59, 999);
-
-  return new Date(currentDate);
-};
+//   return new Date(currentDate);
+// };
 
 export const isFilmExpired = (dueDate) => {
   if (dueDate === null) {
     return false;
   }
-
-  const currentDate = getCurrentDate();
+  return true;
+  // const currentDate = getCurrentDate();
 
   // return currentDate.getTime() > dueDate.getTime();
 };
@@ -21,8 +19,8 @@ export const isFilmExpiringToday = (dueDate) => {
   if (dueDate === null) {
     return false;
   }
-
-  const currentDate = getCurrentDate();
+  return true;
+  // const currentDate = getCurrentDate();
 
   // return currentDate.getTime() === dueDate.getTime();
 };
