+++ src/view/abstract.js
@@ -5,8 +5,8 @@ export default class Abstract {
     if (new.target === Abstract) {
       throw new Error(`Can't instantiate Abstract, only concrete one.`);
     }
-  this._element = null;
-  this._callback = {};
+    this._element = null;
+    this._callback = {};
   }
 
   getTemplate() {
